refactor(mentor): move modules from functions to utils
