+++ utils/moveFolderPage.js
@@ -1,11 +1,11 @@
 import { emailMsg, passwordMsg, emailInput, passwordInput, passwordCheck, passwordCheckMsg } from "../elements.js";
 import { EMAIL, PASSWORD } from "../auth.js";
 import { changeInputColor } from "./changeInputColor.js";
-import { isValidEmail, isValidPassword } from "../utils/validator.js";
+import { isValidEmail, isValidPassword } from "./validator.js";
 import { isUseAlready } from "./isUseAlready.js";
 import { isPasswordCorrect } from "./isPasswordCorrect.js";
-import { ERROR_MESSAGES } from "../utils/messages.js";
-import { redirectToFolderPage } from "../utils/redirectToFolderPage.js";
+import { ERROR_MESSAGES } from "./messages.js";
+import { redirectToFolderPage } from "./redirectToFolderPage.js";
 
 /**
  * 에러메시지를 노출하고 싶은 Element와 노출하고 싶은 메시지를 받아서 단순 UI를 변경해주는 함수
@@ -47,9 +47,7 @@ export const moveFolderSignup = (e) => {
 
   showErrorMessage(emailMsg, isEmailValid ? "" : ERROR_MESSAGES.emptyEmail);
   showErrorMessage(passwordMsg, isPasswordValid ? "" : ERROR_MESSAGES.invalidPassword);
-  showErrorMessage(emailMsg, isPasswordCheck ? "" : ERROR_MESSAGES.incorrectPassword);
-  // emailMsg.innerHTML = !isEmailValid ? "이메일을 확인해 주세요." : "";
-  // passwordMsg.innerHTML = !isPasswordValid ? "비밀번호는 영문, 숫자 조합 8자 이상 입력해 주세요." : "";
-  // !isPasswordCheck ? passwordCheckMsg.classList.remove("hidden") : passwordCheckMsg.classList.add("hidden");
+  showErrorMessage(passwordCheckMsg, isPasswordCheck ? "" : ERROR_MESSAGES.incorrectPassword);
+
   changeInputColor();
 };
