+++ utils/showErrorMsg.js
@@ -1,9 +1,9 @@
 import { changeInputColor } from "./changeInputColor.js";
 import { emailMsg, passwordMsg, passwordCheckMsg, passwordInput } from "../elements.js";
-import { isValidEmail, isValidPassword } from "../utils/validator.js";
+import { isValidEmail, isValidPassword } from "./validator.js";
 import { isUseAlready } from "./isUseAlready.js";
 import { isPasswordCorrect } from "./isPasswordCorrect.js";
-import { ERROR_MESSAGES } from "../utils/messages.js";
+import { ERROR_MESSAGES } from "./messages.js";
 
 const signin = {
   getEmailError(value) {
@@ -16,7 +16,7 @@ const signin = {
   },
 };
 
-const singup = {
+const signup = {
   getEmailError(value) {
     if (!value) return ERROR_MESSAGES.emptyEmail;
     if (!isValidEmail(value)) return ERROR_MESSAGES.invalidEmail;
@@ -35,26 +35,30 @@ const singup = {
 
 export const showErrorMsgSignin = ({ target }) => {
   const { value, classList } = target;
+
   let errorMsg = "";
   if (classList.contains("input-email")) {
     errorMsg = signin.getEmailError(value);
   } else if (classList.contains("input-pw")) {
     errorMsg = signin.getPasswordError(value);
   }
+
   const messageElement = classList.contains("input-email") ? emailMsg : passwordMsg;
+
   messageElement.innerHTML = errorMsg;
   changeInputColor(messageElement, target);
 };
 
 export const showErrorMsgSignup = ({ target }) => {
   const { value, classList } = target;
+
   let errorMsg = "";
   if (classList.contains("input-email")) {
-    errorMsg = singup.getEmailError(value);
+    errorMsg = signup.getEmailError(value);
   } else if (classList.contains("input-pw")) {
-    errorMsg = singup.getPasswordError(value);
+    errorMsg = signup.getPasswordError(value);
   } else if (classList.contains("input-pw-check")) {
-    errorMsg = singup.getPasswordIncorrectError(value);
+    errorMsg = signup.getPasswordIncorrectError(value);
   }
 
   let messageElement;
